# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the AI Resume Screening & Candidate Matching research prototype.
 * OpenAPI spec version: 0.1.0
 */
export interface Error {
  error: string;
}

export interface HealthStatus {
  status: string;
  mode?: string;
}

export interface DashboardStats {
  totalResumes: number;
  screenedCandidates: number;
  shortlistedCandidates: number;
  averageMatch: number;
}

export interface ChartPoint {
  label: string;
  value: number;
}

export interface CandidateFairness {
  demographicsAvailable: boolean;
  message: string;
}

export interface Candidate {
  id: string;
  name: string;
  email: string;
  phone: string;
  skills: string[];
  education: string;
  experience: number;
  jobTitles?: string[];
  certifications: string[];
  semanticScore: number;
  mlScore: number;
  overallScore: number;
  status: string;
  matchedSkills: string[];
  missingSkills: string[];
  relatedSkills: string[];
  explanation: string;
  fairness: CandidateFairness;
  notes: string;
  sourceFile: string;
}

export interface Dashboard {
  stats: DashboardStats;
  scoreDistribution: ChartPoint[];
  statusBreakdown: ChartPoint[];
  topSkills: ChartPoint[];
  recentScreenings: Candidate[];
  mode: string;
}

export interface Job {
  id: string;
  title: string;
  department: string;
  requiredSkills: string[];
  preferredSkills: string[];
  education: string;
  minimumExperience: number;
  certifications: string[];
  description: string;
  createdAt: string;
  updatedAt: string;
}

export interface JobInput {
  /** @minLength 1 */
  title: string;
  department: string;
  requiredSkills: string[];
  preferredSkills: string[];
  education: string;
  /** @minimum 0 */
  minimumExperience: number;
  certifications: string[];
  description: string;
}

export type JobUpdate = JobInput;

export interface ResumeInput {
  filename: string;
  text: string;
  fileType?: string;
}

export interface ResumeUploadInput {
  resumes: ResumeInput[];
}

export interface Resume {
  id: string;
  filename: string;
  fileType: string;
  status: string;
  uploadedAt: string;
}

export interface ScreeningInput {
  /** @nullable */
  jobId?: string | null;
  /** @minLength 1 */
  jobDescription: string;
  /** @minItems 1 */
  resumes: ResumeInput[];
}

export interface Screening {
  id: string;
  jobTitle: string;
  screenedAt: string;
  results: Candidate[];
  mode: string;
  pipeline: string[];
}

export interface ScreeningSummary {
  id: string;
  jobTitle: string;
  screenedAt: string;
  resumeCount: number;
  shortlistedCount: number;
  averageScore: number;
}

export interface CandidateUpdate {
  status?: string;
  notes?: string;
}

export interface ExplainInput {
  candidateId: string;
}

export interface FeatureImpact {
  feature: string;
  impact: number;
  direction: string;
}

export interface Explanation {
  candidateId: string;
  features: FeatureImpact[];
  interpretation: string;
  mode: string;
}

export interface FairnessMetric {
  name: string;
  value: number;
  interpretation: string;
}

export interface FairnessGroup {
  label: string;
  selectionRate: number;
}

export interface Fairness {
  demographicsAvailable: boolean;
  message: string;
  metrics: FairnessMetric[];
  groups: FairnessGroup[];
  mode: string;
}

export interface Metric {
  name: string;
  value: number;
}

export interface ModelComparison {
  model: string;
  f1: number;
}

export interface ModelMetrics {
  mode: string;
  metrics: Metric[];
  confusionMatrix: number[][];
  comparison: ModelComparison[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

export type ListCandidatesParams = {
search?: string;
status?: string;
skill?: string;
};