# packages/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the AI Resume Screening & Candidate Matching research prototype.
 * OpenAPI spec version: 0.1.0
 */
import type { Candidate } from './candidate';
import type { ChartPoint } from './chartPoint';
import type { DashboardStats } from './dashboardStats';

export interface Dashboard {
  stats: DashboardStats;
  scoreDistribution: ChartPoint[];
  statusBreakdown: ChartPoint[];
  topSkills: ChartPoint[];
  recentScreenings: Candidate[];
  mode: string;
}