# packages/api-zod/src/generated/types/explanation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the AI Resume Screening & Candidate Matching research prototype.
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureImpact } from './featureImpact';

export interface Explanation {
  candidateId: string;
  features: FeatureImpact[];
  interpretation: string;
  mode: string;
}