# packages/api-zod/src/generated/types/fairness.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the AI Resume Screening & Candidate Matching research prototype.
 * OpenAPI spec version: 0.1.0
 */
import type { FairnessGroup } from './fairnessGroup';
import type { FairnessMetric } from './fairnessMetric';

export interface Fairness {
  demographicsAvailable: boolean;
  message: string;
  metrics: FairnessMetric[];
  groups: FairnessGroup[];
  mode: string;
}