# packages/api-zod/src/generated/types/modelMetrics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the AI Resume Screening & Candidate Matching research prototype.
 * OpenAPI spec version: 0.1.0
 */
import type { Metric } from './metric';
import type { ModelComparison } from './modelComparison';

export interface ModelMetrics {
  mode: string;
  metrics: Metric[];
  confusionMatrix: number[][];
  comparison: ModelComparison[];
}