# packages/api-zod/src/generated/types/resumeUploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the AI Resume Screening & Candidate Matching research prototype.
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeInput } from './resumeInput';

export interface ResumeUploadInput {
  resumes: ResumeInput[];
}