# packages/api-zod/src/generated/types/screeningSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the AI Resume Screening & Candidate Matching research prototype.
 * OpenAPI spec version: 0.1.0
 */

export interface ScreeningSummary {
  id: string;
  jobTitle: string;
  screenedAt: string;
  resumeCount: number;
  shortlistedCount: number;
  averageScore: number;
}